Closes #108.
Confirmed that clicking the close button in some browsers (FireFox, IE10) was not closing the toast. The problem was that the focus was still on the toast, and we want to toast to hang around normally when it has focus. So we need the focus code. But when we click the close button, now we explictly tell it to ignore the focus code.