Improvements to make sass BS4 compatible.
This change updates the SASS file so that it can be used better with BS4 by aligning to BS4 breakpoints and using the BS4 variables for the colors (keeping the colors consistent and making it easier to change the branding to match the BS4 branding colors).  This change will still work if the BS4 source is not included by having all the required variables and mixins right in this source file, however, if a dev has the BS4 source, uncomment the import statements and the BS4 source will override the values defined in this file.

Changes to this file also were done to use a 'mobile-first' approach of BS4.  Therefore some of the breakpoints might not look exactly as they did before this change, but I did some quick testing at smaller screen sizes and it looked ok.  There are also some things that were not tested such as the 'close button' and other options that should be double checked before this is used in the main branch.